Playlist struct, Backend trait, restructuring code
